+++ src/Select/Select.component.tsx
@@ -178,6 +178,12 @@ export interface Props extends React.HTMLAttributes<HTMLDivElement> {
    * @default ''
    */
   menuPortalTarget?: HTMLElement;
+  /**
+   * Is the filled element highlighted
+   *
+   * @default false
+   **/
+  highlightFilled?: boolean;
 }
 
 const SDiv = styled.div<Props>`
@@ -219,7 +225,8 @@ const SDiv = styled.div<Props>`
    }    
   .react-select-component {
     .react-select__control {
-      min-height: auto;
+      min-height: ${(props: Props) =>
+        props.theme.select[props.selectSize!].height};
       border-radius: ${(props: Props) =>
         props.theme.select.borderRadius ||
         props.theme.common[props.selectSize!].borderRadius};
@@ -294,6 +301,9 @@ const SDiv = styled.div<Props>`
       }
     }
     
+    &.highlighted .react-select__control {
+      background-color: ${props => props.theme.colors.highlight200};
+    }
   }  
 `;
 
@@ -332,6 +342,11 @@ export class CustomSelect extends React.Component<Props> {
         ? this.props.closeMenuOnSelect
         : !isMulti;
     const uniqueId = uuid.v4();
+
+    const selectValue = this.props.value || this.props.selectedOption;
+    const valueIsNotEmpty: boolean =
+      !!selectValue && (!Array.isArray(selectValue) || selectValue.length > 0);
+
     return (
       <ThemeProvider theme={(outerTheme: any) => outerTheme || theme}>
         <SDiv
@@ -344,7 +359,6 @@ export class CustomSelect extends React.Component<Props> {
         >
           <BaseSelectComponent
             closeMenuOnSelect={closeMenuOnSelect}
-            className="react-select-component"
             classNamePrefix="react-select"
             isDisabled={isDisabled}
             isClearable={isClearable}
@@ -360,6 +374,9 @@ export class CustomSelect extends React.Component<Props> {
             menuPlacement={'bottom'}
             {...props}
             {...controlSpecificProps}
+            className={`react-select-component ${
+              this.props.highlightFilled && valueIsNotEmpty ? 'highlighted' : ''
+            } ${props.className}`}
           />
           {invalid && (
             <ErrorMessage